+++ src/libtrx/include/libtrx/config/types_tr1.h
@@ -201,7 +201,8 @@ typedef struct {
         bool enable_perspective_filter;
         GFX_TEXTURE_FILTER texture_filter;
         GFX_TEXTURE_FILTER fbo_filter;
-        bool enable_debug;
+        bool enable_debug_triggers;
+        bool enable_debug_portals;
         bool enable_wireframe;
         double wireframe_width;
         bool enable_vsync;