+++ src/tr1/game/output.h
@@ -38,11 +38,14 @@ bool Output_IsSkyboxEnabled(void);
 void Output_DrawSkybox(const OBJECT_MESH *mesh);
 
 void Output_DrawRoom(const ROOM_MESH *mesh);
+void Output_DrawRoomPortals(const ROOM *room);
 void Output_DrawRoomTriggers(const ROOM *room);
 void Output_DrawShadow(int16_t size, const BOUNDS_16 *bounds, const ITEM *item);
 void Output_DrawLightningSegment(
     int32_t x1, int32_t y1, int32_t z1, int32_t x2, int32_t y2, int32_t z2,
     int32_t width);
+void Output_Draw3DLine(XYZ_32 pos_0, XYZ_32 pos_1, RGBA_8888 color);
+void Output_Draw3DFrame(const XYZ_32 vert[4], RGBA_8888 color);
 void Output_FlushTranslucentObjects(void);
 
 void Output_DrawScreenFlatQuad(