+++ src/tr1/specific/s_output.h
@@ -45,6 +45,8 @@ void S_Output_DrawTexturedTriangle(
 void S_Output_DrawTexturedQuad(
     PHD_VBUF *vn1, PHD_VBUF *vn2, PHD_VBUF *vn3, PHD_VBUF *vn4, int16_t tpage,
     uint16_t textype);
+void S_Output_Draw3DLine(
+    const PHD_VBUF *vn1, const PHD_VBUF *vn2, const RGBA_8888 color);
 void S_Output_DrawSprite(
     int16_t x1, int16_t y1, int16_t x2, int y2, int z, int sprnum, int shade);
 void S_Output_Draw2DLine(