+++ docs/tr1/CHANGELOG.md
@@ -16,7 +16,7 @@
 - added a `/cut` (alias: `/cutscene`) console command for playing cutscenes
 - added a `/gym` (alias: `/home`) console command for playing Lara's Home
 - added a `/music` console command that plays a specific music track
-- added a `/debug` console command that shows all triggers in pink
+- added a `/debug` console command that shows all triggers and portals
 - added a console log when using the `/demo` command
 - ⚠️ changed the game data to use a separate strings file for text information, removing it from the game flow file
 - ⚠️ changed the game flow file internal structure